Extract small classes to files
